Move RobotFactory to utils
